feat(scoped-elements): self-registering components compatibility
Reusing the global tag name can be useful in some situations in which not all component libraries are exporting the component classes or when components are looking in DOM by specific tag names. In those cases, although this change doesn't remove the limitation of having just one version of those elements in our app, it makes them work properly inside our scoped elements as we are going to use the tag name they use in their definition.