fix(create): generate sourcemap for ts files
Generates a separate .js.map file in TS `outDir` with the original TS source inlined in it.

This should let us see the original TS source while debugging.

Browser is able to load the sourcemap when used in development mode with es-dev-server.
Currently not being processed by rollup in production mode and the behavior is same as before.