feat(eslint-config): upgrade eslint to 7.x
BREAKING CHANGE: for a guide to migrate your code to eslint 7.x, as well as a list of breaking changes, please refer to:
- https://eslint.org/docs/user-guide/migrating-to-6.0.0
- https://eslint.org/docs/user-guide/migrating-to-7.0.0

Additionally, this change also removes the usage of `babel-eslint`. For most projects, this should work transparently. However, if your code specifically references one of its rules, you can:
- add `@babel/eslint-parser` to your devDependencies
or
- remove the references, as some of the rules have been merged into eslint