Adding Application publishing link to developing-components/publishing.md
Component publishing is surfaced before application and when looking for "publishing" or "build" info. Hence I assumed that build is not really applicable in open-wc overall. Which is obviously not true as it is in scope of application publishing. The cross-reference on different approaches to publishing in Component and in Application docs would clear the difference. 
Unfortunately for application I have not found similar page, so used rollup build page for this reference.