fix: scoped registries polyfill
It could happen that the polyfill is not applied in the document header but on demand when needed. That could make some classes to have the native HTMLElement in its hierarchy instead of the patched HTMLElement of the polyfill, which makes the instantiation of the component to fail. This commit checks the hierarchy of the class to be instantiated and patches the HTMLElement if needed.