add Svelte to list of base librarryies
There was always an option to build custom elements with svelte, but recently there was a [huge update](https://github.com/sveltejs/svelte/pull/8457) and now dev experience is much better that was before. Lots of issues were fixed and now svelte is a pretty solid solution to build reliable custom elements. It would be nice to add it to the list of base libraries.

For example lots of svelte libraries were complied to custom elements before, (slidy)[https://github.com/Valexr/slidy/tree/main/packages] for example, and now it's even much easier than befor.