fix(eslint-plugin-lit-a11y) fix improper regex
The existing regex would prevent valid custom elements
from being detected appropriately.

The [spec](https://html.spec.whatwg.org/multipage/custom-elements.html#valid-custom-element-name)
indicates that a valid custom name must start with a lowercase a-z but
the next word between that character and the hyphen doesn't necessarily
need to be present as the `+` in the existing regex indicates.

This makes it so custom elements such as `<d-button>` or `<q-tooltip>` are
detected as such.