Improved example on web-dev-server.config
The previous example called a function but didn't show where we were expect to insert that call. It also had a semicolon, which shouldn't be there, because we are supposed to insert the function call in the middle of the configuration object.

This new example is probably a bit clearer and easier to understand. It is based on https://www.npmjs.com/package/@web/dev-server-polyfill

Be aware, though, I'm not sure this new example is correct. Someone please double-check and improve as needed.