chore: added patch-package to dependencies
When first setting up this repository, at the end of `npm ci` it asks us to
install `patch-package`. That's odd, such command should "just work" without any
extra step.

The solution is simple: add it to the `dependencies` in the `package.json`.
Since this is required for the proper setup of this repository, it makes sense
to be listed as a dependency.