fix(eslint-plugin-lit-a11y) support all tags in allowList
This change resolves an issue where the `allowList` was only being applied to custom-elements. As far as I can see, the `allowList` option is intended to let users “_specify tag names to exclude from this rule_” i.e all tag names and not _just_ custom-elements tag names.

Supporting all tag names in the `allowList` is particularly valuable for `<slot>`, where a design system might wish to intercept clicks on user-defined buttons allowed in a `<slot @click=${this.handleClick}>`.